Move the changelog to the root directory
As we don’t want to have separate versions for each platform, we don’t
need a platform-specific changelog either.  This patch moves the
changelog from the runners/lpc55 directory to the root directory to make
it easier to discover and to make it clear that it applies for all
targets, not only for the lpc55.